D-Lite-call-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-socketio==5.13.0

